Fix build.

+++ lib/sources/components/kernel.cpp
@@ -227,7 +227,7 @@ containers::async_of<shared<event>> kernel::handle(uuid session_id, uuid websock
         _event->data_ = {
             {"command", "ack"},
             {"message", "The given data be a valid JSON."},
-            {"status", _event->status_code_},
+            {"status", static_cast<int>(_event->status_code_)},
         };
         co_return _event;
     }
@@ -244,7 +244,7 @@ containers::async_of<shared<event>> kernel::handle(uuid session_id, uuid websock
             {"command", "ack"},
             {"message", "The given data was invalid."},
             {"errors", _validator->errors_},
-            {"status", _event->status_code_},
+            {"status", static_cast<int>(_event->status_code_)},
         };
 
         co_return _event;
